Rename app to webapp and add cli and library scaffolding
