+++ README.md
@@ -6,7 +6,7 @@ pip install -r requirements.txt
 古いバージョンの pip を使っている場合は、先にアップデート
 python -m pip install --upgrade pip
 
-pybabel compile -d app/translations -f
+pybabel compile -d webapp/translations -f
