+++ cli/main.py
@@ -1,0 +1,11 @@
+import click
+from mylib import utils
+
+@click.command()
+@click.argument('name', default='World')
+def main(name: str) -> None:
+    """Simple CLI entrypoint."""
+    click.echo(utils.greet(name))
+
+if __name__ == '__main__':
+    main()
