+++ mylib/__init__.py
@@ -1,0 +1,5 @@
+"""Common library package."""
+
+from .utils import greet
+
+__all__ = ["greet"]
