+++ mylib/utils.py
@@ -1,0 +1,3 @@
+def greet(name: str) -> str:
+    """Return a friendly greeting."""
+    return f"Hello {name}"
