+++ requirements.txt
@@ -10,3 +10,4 @@ qrcode
 Pillow
 requests
 cryptography
+click
