+++ webapp/auth/templates/auth/register_no_totp.html
