+++ webapp/feature_x/templates/feature_x/dashboard.html
