+++ webapp/photo_view/routes.py
