+++ wsgi.py
@@ -1,4 +1,4 @@
-from app import create_app
+from webapp import create_app
 app = create_app()
 
 if __name__ == "__main__":
